requirements.txt
python-telegram-bot==21.9
openai==1.61.0
